Use gcThreadCountSpecified
Make use of newly introduced gcThreadCountSpecified flag so that when
command line options (such as Java's -Xgcthreads or -Xgcmaxthreads) are
specified, GC thread count overrides the default count calculation.

Old gcThreadCountForced flag is still used, butin ParallelDispatcher to
distinguish between force/fixed or adaptive count of threads, bufore
deciding how many threads to use for a ParallelTask.

Another change is that during the snapshot VM creation when thread count
is reduced to checkpointGCthreadCount, we don't adjust GCExtensions
gcThreadCount anymore (only ParallelDispatcher is affected). This is to
be able to carry over the original count (from snapshot VM) to restore
VM, so that options that specify that count can be obeyed even on
restore side.