(0.42) AArch64: Add class unload pic to class constant under ifacmp node
Class unload pics must be added to class constants so that constants can be
invalidated when they are unloaded. aconstEvaluator does add them on AArch64.
However, ificmpHelper does not evaluate aconst node if the constant value can be
encoded into the immediate field of the cmp instruction.
This commit updates ificmpHelper to evaluate aconst node if class unload pic
is required.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>