Add flag to indicate compile time is not that important
This commit adds a new option bit called TR_NotCompileTimeSensitive
which will be turned on by an upstream project (OpenJ9) in situations
where we favor throughput over compile time.
The commit also adds a new optimization qualifier, `IfLoopsAndNotCompileTimeSensitive`.
An optimization pass flagged as such will be performed only if
the method being compiled has loops and the `TR_NotCompileTimeSensitive`
bit is turned on.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>