GRA changes to reduce JIT overhead at warm opt levels
The commit introduces a new option -Xjit:graFreqThresholdAtWarm=<NNN>
that can be used to control the JIT overhead for compilations
performed at warm optimization level or below. The default value
of the threshold is 500. A larger value leads to lower JIT overhead
but also to potentially lower quality of the generated code.
For hot compilations or in situations where we think we can afford
more JIT overhead, the threshold is automatically set to 0.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>