Facilitate recording of constants observed in call target selection
If inlining uses constant folding to help determine call targets, then
it's important for the corresponding IL (if any is generated) to be
consistent with what the inliner saw.

This commit equips TR_CallTarget with a collection of observed constants
that must be folded consistently in IL. This collection can be populated
by an inliner and later consulted by an IL generator to ensure that the
operations are folded as needed.

This repeated folding is important whenever a value might be allowed to
be folded despite the possibility of a later change. It also allows
constants to be speculative by specifying the assumptions that are
necessary in order for the folding to be correct, and by informing the
IL generator of the locations where such assumptions have been made.