RISC-V: fix docker image creation for CI cross-builds
This commit fixes Dockerfile creating image for RISC-V CI builds. The
problem was that URL libfakeroot and libfakechroot was no longer valid.
It pointed to Debian ports archive, but meanwhile RISC-V architecture
became "officially" supported (will be once Trixie is out).

It is not clear to me why Dockerfile downloaded and extracted RISC-V
versions of libfakeroot and libfakechroot at all. However, I had been
successfully building sysroots for OMR and/or OpenJ9 compilations using

    mmdebstrap --mode=fakeroot --variant=extract

without need of RISC-V libfakeroot and libfakechroot. So this commit
removes download and manual installation of them and uses
--variant=extract to build sysroot. In addition it uses new ("official")
Debian archives to build the sysroot. This way the aforementioned
problem with outdated URLs should be fixed for good.

Fixes #7256.