Infra for mmap memory backed by temporary files
This commit adds support for allocating memory using
mmap and backing that memory by a temporary file.
The goal is to reduce the amount of physical memory in use
by saving the content of an infrequently used memory segment
to its backing file using the madvise() system call with
a hint of MADV_PAGEOUT.
This feature is only available for Linux.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>