Remove flushCachesForGC from acquireExclusiveForGC path
flushCachesForGC is mostly called close/prior to start-gc point for each
particual GC operation/STW-increment, so that we often end up doing it
twice, the first one being early during acquireExclusiveVMAccessForGC
negotiation.

This change removes the one from acquireExclusiveVMAccessForGC, but also
adds a couple of flushCachesForGC in GC specific operation that relied
on acquireExclusiveVMAccessForGC to do it.

Signed-off-by: Aleksandar Micic <Aleksandar_Micic@ca.ibm.com>