Fix typo AIXPPCX -> AIXPPC
On AIX, this typo disables code that prevents upcall_handler from
collecting thread data when upcall_handler is unregistered. Collecting
data after upcall_handler is unregistered might cause corrupt memory
to be passed through siginfo and lead to crashes.

More context:
On AIX, a signal can can be unavailable to sigpending/sigwait but
not yet have been delivered to a thread. As a result, it is not
guaranteed that SIG_SUSPEND won't be delivered after upcall_handler
is unregistered.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>