Upgrade pugixml 1.5 to 1.8
pugixml 1.8 was chosen because it can still be built using the same build files
as pugixml 1.5.

Miscellanous modifications to pugixml.cpp were made for EBCDIC support and for
snprintf compatibility with modern build compilers.

pugixml 1.5 is deleted from OMR.

The main motivation for this upgrade is to simplify the license vetting process
with the Eclipse Foundation.  However, it is also useful to maintain currency
with newer pugixml releases.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>