Create debug counters correctly for removed call sites
This was passing callsite->_callNodeTreeTop while it was still null,
which prevented the debug counter increment from being generated and
even prevented a static debug counter increment.

The treetop where the call occurs is tt, and in this case callsite is
the result of searching for a dead call site that matches it. This tt
is already used for the Unresolved_In_ECS case, and it's also the value
that callsite->_callNodeTreeTop would take on just below if callsite
hadn't been removed.