Update Array helper APIs in ValuePropogationCommon for OffHeap
This commit is based on ed01b6dc36e7362986eda898463cc960135a82a5
Instead of modifying the code to support off-heap which as-is breaks
runs with arraylet environments, this commit adds separate paths for
off-heap while keeping the original paths as-is.

Code can be modified similar to the based-on commit to have gencon
runs using the array helper APIs while not affecting arraylet runs.

Signed-off-by: Abdulrahman Alattas <rmnattas@gmail.com>