Open XL segfault fix and workaround for si.cpp macro collision
The omrthread_get_ras_tid() function in rasthrsup.c was causing
a segmentation fault with Open XL as pthread_t is structured in a
different manner. This fixes that while keeping support with XLC.
The header grp.h used in si.cpp also has a macro collision with the
native locale header from Open XL. The reordering of headers is a
workaround for the core problem. A issue has also been created in
order to use a better macro name __grp.