Use testIsClassArrayType to generate IL to test for array type
The downstream OpenJ9 project defines a new method,
TR_VMBase::testIsClassArrayType, that generates IL to tests whether a
class is an array class.  Change code that generated IL that loaded
classDepthAndFlags field itself to call testIsClassArrayType instead to
hide at least some of the J9-specific details.

Also, generate a comparison of the result of the IL produced by
testIsClassArrayType with zero, rather than comparing with the array
class flag.  A comparison with zero is expected to be preferred on most
platforms.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>