Correct typos mentioning classDepthAndFlags field
Code in OMR supports the JIT compiler in generating IL to access the
J9Class field classDepthAndFlags field in the downstream OpenJ9 project.
In many cases, that code incorrectly refers to the field as
classAndDepthFlags.

This commit adds new duplicate definitions of the methods and enum
values that correct the name.  Once the downstream project removes
references to the obsolete methods, the obsolete definitions will be
removed.