Kill prior store nodes in processing volatile store in local transparency
When processing a volatile store in `TR_LocalTransparency::updateUsesAndDefs`,
it should kill any store node that has appeared earlier in the block.
Otherwise, the store node could survive the block in local transparency
but will be killed in `_downwardExposedAnalysisInfo` when the volatile
store is processed in local anticipatability, which causes
inconsistency between local transparency and local anticipatability.
This matches how a call node is processed in
`TR_LocalTransparency::updateUsesAndDefs`.

Fixes: eclipse-openj9/openj9#19218

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>