Delete TR::SelectOpt, which is unused
If there is a case in which it's desirable for an entry in _opts to
select between two different Optimization classes based on an option, we
can simply check the option where that entry is initialized, and use it
to choose which factory function to pass to the OptimizationManager.
Moreover, by checking the option there, we'll naturally take option sets
into account.