Tweak a2e header wrappers for Open XL compliance
When compiling with omr_ascii option, there is an alternative notation
to pass in the a2e header wrappers with Open XL using the -isystem
option. While making this work, it was also suggested to cleanup the
PREFIXPATH(header) notation to simply use #include_next which supports
and works with both XLC and Open XL on z/OS.

dlfcn.h was added as prerequisite for the dll.h a2e header wrapper.

Signed-off-by: Gaurav Chaudhari <gaurav.chaudhari@ibm.com>