Initialize the array variable in generateArrayAddressTree to null
The code populates the `TR::Node* array` variable if a non-zero offset exist.

The variable is passed to `generateArrayElementAddressTrees` which
depends on if the variable is null to indicate a zero offset.

Given that the variable is not initialized to null a value may be passed
causing wrong generated trees.

Signed-off-by: Abdulrahman Alattas <rmnattas@gmail.com>