Fix -Wnonnull warnings with GCC 11
Version 11 and later of GCC consider the implicit `this` parameter of
nonstatic member functions to be annotated with `nonnull` for the
purposes of the `-Wnonnull` warning. This change generates new warnings
that are resolved here with the use of `TR_ASSERT_FATAL` on the objects
that the compiler thinks could be `NULL`. The macro `TR_ASSERT` is
insufficient, as the `TR::assertion` method isn't (and cannot be)
annotated with `OMR_NORETURN`.

Signed-off-by: Christian Despres <despresc@ibm.com>