Prevent VGHM from moving guards in the presence of shadow symrefs
In some cases the VGHM opt was moving code above a runtime guard allowing a privatized argument load to read past the end of the object bounds. This privatized temp slot was marked collectible causing the GC to assert when the loaded value was not a valid object reference. The issue the assert was reporting would not cause any runtime behaviour issues as the guard would have prevented any of the inlined code from executing and therefore the temp would never have been used, but allowing the invalid load to happen could cause a SEGV if the load offset caused an access past the end of the heap. This fix looks for shadow symrefs between the two guards and aborts guard movement when one is found.

OMR PR: https://github.com/eclipse-omr/omr/pull/7411