Correctly handle isInstanceOf()==TR_maybe in VPClass::mustBeNotEqual()
Both fixedClass and boundClass are known at compile time, so usually the
result of isInstanceOf() can be only TR_yes or TR_no. However, in AOT
without SVM we can't be allowed to know the subtyping relationship, so
the result is instead TR_maybe. This logic was ignoring the possibility
of TR_maybe and treating it in the same way as TR_no, which indicates
that the types are unrelated and the values must be distinct, but that's
incorrect. Now instead it will conservatively treat TR_maybe the same
way it treats TR_yes, i.e. as an indication that the types are related
and therefore the values could be equal. That result is still safe even
if the types are actually unrelated.