Fix error: comparison of unsigned expression < 0 is always false
omrthread_prio_t is of type uintptr_t. It cannot have a value below 0.

J9THREAD_PRIORITY_MIN is defined as 0 and J9THREAD_PRIORITY_MAX is
defined as 11.

In omrthread_attr_set_priority, priority, which is of type
omrthread_prio_t, will always be greater than or equal to
J9THREAD_PRIORITY_MIN. To see if priority is invalid, we only need
to evaluate if it is greater than J9THREAD_PRIORITY_MAX.

Fixes: #7454

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>