Fix blocksplitter for trivial conditional branches
Certain conditional branches can be made trivial by earlier opts. For
these the branch target may be the fall-through path. If so, since we're
redirecting the fall-through path to a goto block and removing the
single edge that represents both the fall-through and branch paths, we
should also redirect the branch to the goto block.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>