Use fallocate instead of ftruncate
Even if the ftruncate() system call succeeds, this does not guarantee
that the underlying filesystem has enough free blocks to support
whatever increase in file size that was requested. This can cause SIGBUS
errors if writes are attempted in the file.

The fallocate() call does guarantee that there is enough space
available.

Signed-off-by: Christian Despres <despresc@ibm.com>