ZOS: Fix const length buffer used in atoe methods
Some of the atoe methods were using a constant length buffer for string
conversion. If the input format string and arguments are larger than the
buffer size, buffer overflow would happen. This bug was fixed by running
atoe_vsnprintf with NULL buffer to get the required buffer size then create
the buffer and write in it.

Signed-off-by: Ehsan Kiani Far <ehsan.kianifar@gmail.com>