Re-enable RegDepCopyRemoval for value types
As documented in https://github.com/eclipse-openj9/openj9/issues/9712,
initial attempt to fix the SplitPostGRA to work with RegDepCopyRemoval
resulted in 5% regression in performance throughput on x86 which
required us to disable the RegDepCopyRemoval when value types are
enabled. Later on the performance issue was addressed and fixed. This PR
reverts the change that disabled RegDepCopyRemoval when value types are
enabled.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>