Add jProfileValue symbols in TR_Debug::getName for symrefs
While looking into the JIT compilation log, I bumped into the case where
bunch of jProfiling Value place holder calls were printed as unknown in
the log, as it was missing from the Debug.cpp. Commit in this PR adds
it.
Also For some of the non-helper call symbols, RAS helper function that
returns name of the symbol, remove the "Symbol" from the name that will
be returned.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>