Move willBeEvaluatedAsCallByCodeGen to OpenJ9
The willBeEvaluatedAsCallByCodeGen implementation for x is almost
entirely guarded by the J9_PROJECT_SPECIFIC macro. If the macro is not
set, it always returns true which is what the default implementation
already does. Since this implementation is only meaningful to OpenJ9, it
is being moved there.

willNotInlineCompareAndSwapNative is only called from this
implementation of willBeEvaluatedAsCallByCodeGen so it is being moved as
well.

Signed-off-by: jimmyk <jimmyk@ca.ibm.com>