Aarch64: Fix isMethodPointerConstant call on wrong node
While testing jb2 on my macbook, I ran into a crash in JIT asking
an ifcmp node if it was a method pointer constant, which should
only be asked of aconsts or loadaddrs. There was an obvious typo
asking this question of "node" rather than "secondChild" so this
fixes that typo.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>