Track distinct live commoned loads of locals for Store Sinking
Store sinking tracks which local variables are used in the value
assigned to a candidate store, tracks whether a load of a local variable
has been commoned, and whether the load might be killed by a store to
that local.  All this tracking is done on the basis of the index of the
local variable.

However, it's possible for a local to be loaded before a store to that
local, and the value that existed before that store to be used by the
candidate for store sinking.  In general, there could be several loads
and stores to such a local interleaved, with each of the distinct values
loaded being used by commoned references in the store that's a candidate
for store sinking.

This problem was fixed by having LiveVariableInformation keep track of
the distinct nodes that load local variables in backwards tree walks,
and the number of distinct loads for each local variable.

Also removed a redundant copy into *savedLiveCommonedLoads of the result
of a call to TR_LiveVariableInformation::liveCommonedLoads(), and
removed an unused TR_BitVector satisfiedLiveCommonedLoads.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>