Move UnsafeSubexpressionRemoval to its own file
UnsafeSubexpressionRemoval is currently used in IsolatedStoreElimination
where it is defined.  The Store Sinking optimization requires the utility
provided by UnsafeSubexpressionRemoval, so this change moves the
declaration and implementation of that class into there own files.

Also, if the reference count of a store that is to be deleted is non-
zero, recreate its parent node - which should be a check or
compressedRefs node - as a TR::treetop.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>