Remove compiler ras address masking
"Address masking" was an attempt to make compiler log files easier to compare by
preventing addresses (that are likely to change between compilations) from being
printed raw in a compilation log.  It worked by intercepting the "printf" family
of IO functions and checking whether any addresses were being printed with the
`%p` format specifier, and instead of printing the address a "*Masked*" string
was substituted instead.

This feature was introduced at a time when text file comparison tools were not
as robust as they are today.  As well, these days, many addresses are output to
the log without using `%p` making this feature less effective.  Since it adds
overhead and complexity to the logging path, has incomplete coverage, and is not
widely used (if at all), remove it from the code base.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>