Fix syncTempTrampolines when startPC returns 0
When syncing temporary trampolines, there is a call to startPC with the
callee method to try and get the start PC. If the latest compilation of
that method failed, startPC returns zero which is not a real start PC
and can not be used as a target for a trampoline. Attempting to do so
will cause an error.

This change adds error checking when trying to sync trampolines. When a
start PC of zero is returned from the call to startPC, the previous
start PC is used instead.

Signed-off-by: jimmyk <jimmyk@ca.ibm.com>