Zero offheap on commit on OSX
For some reason OSX does not seem to zero memory on commit, unless it's
the first commit. For now, till we understand if there is a way to force
zero on subsequent re-commits, we will zero it explicitly.

Including zPTF as well, since we don't know how it will behave, even
if/when it adds support for commit/decommit/virtual-memory.

Signed-off-by: Aleksandar Micic <Aleksandar_Micic@ca.ibm.com>