Add condition for when SIMD instructions are used for fmin/fmax on Z
SIMD instructions generated for fmax/fmin operations (Short floating
point format) are supported with Vector Enhancement Facility 1 which
became available from IBM Z14 and newer.
Changes in this PR adds helpers for fmin/fmax/dmin/dmax that check that
SIMD instructions for the given type are supported on the platform.

Signed-off-by: Matthew Hall <matthew.hall3@outlook.com>