Remember whether dltOptLevel= is specified in a subset
The command line option -Xjit:dltOptLevel=... existed,
but it was inspected in the OpenJ9 downstream project
only when the `TR_DebugDLT` env var was set. The reason
is that we have to account for the possibility of specifying
the option in a subset and retrieving options from
subsets is relatively expensive and not something that we
may want to do too often.
This commit remembers whether `dltOptLevel=` is specified
in any subset by setting a flag in the JIT command line options.
Then, in OpenJ9 we can first check if this flag is set and
only then proceed to find the correct subset for the method
in question. This should eliminate the overhead when
-Xjit:dltOptLevel= is not specified in a subset, which is
the common case.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>