Move mask overlap check from omrgcconsts.h to ForwardedHeader.hpp
The latter (a C++ header file) cannot be included in the former (which
must be usable in C source files), and it's not reasonable to assume
that the latter is always included before the former.

ForwardedHeader.hpp is included in ForwardedHeader.cpp (and other
compilation units) so the check is guaranteed to be done.

Add missing parentheses.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>