Define jitThrowIdentityException helper and isIdentityObject non-helper
The <isIdentityObject> non-helper can be used by downstream projects as
a way of distinguishing in IL whether an object is an instance of a
class for which each instance has a unique identity, sometimes called an
identity type, versus an instance of a class for which instances that
have the same value are indistinguishable, sometimes called a value
type.

The <jitThrowIdentityException> helper can be used by downstream
projects in IL to throw an IdentityException.  It would typically be
used to report that an object was expected to be an instance of a class
for which each instance has a unique identity.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>