PRE fix to handle data address pointer correctly
Data address pointers are internal pointers and need special
handling in the PRE optimization. Like aladd/aiadd, they are
skipped over in the analysis and optimization itself, but
we are aiming to allow parent nodes that contain such a node
as a child to participate in the analysis and get optimized.
This commit changes the local analyses in PRE to treat data
address pointer analogously to how aladd/aiadd.

Signed-off-by: Vijay Sundaresan <vijaysun@ca.ibm.com>