Enabled noinline workaround for functions in omrzfs.c
Some functions containing the native call BPX4PCT are causing
a crash with Open XL, when they are called back to back in omrzfs. Instead
of lowering optimization level to O0 entirely, specifying
no inlining for these functions seems to mitigate this problem.
This will be used as a workaround until a permanent fix is
released.

Signed-off-by: Gaurav Chaudhari <gaurav.chaudhari@ibm.com>