Reorder J9ZOS390 conditionals in AtomicSupport.hpp for Open XL z/OS
When Open XL AIX support was merged earlier, it seems to have
caused problems with Open XL z/OS builds. The __open_xl__
conditionals meant for AIX are taking priority over the
J9ZOS390 clauses, causing problems for z/OS. Was recommended
to reorder these so that the more granular J9ZOS390 conditional
is encountered and entered first.
The old volatile uint32_t workaround seems to be no longer valid
as the minimal zos version is 2.3, and is removed after testing
it works without the workaround.

Signed-off-by: Gaurav Chaudhari <gaurav.chaudhari@ibm.com>