Only free options objects in Options::shutdown if safe for FrontEnd
Adds a query to TR_FrontEnd (which defaults to "not safe") to answer
whether Options::shutdown() should free the command-line options
objects. For compilers like JitBuilder and the Test compiler, not
freeing the options represents a memory leak and there is not reason
not to free the options objects, so these compilers arrange for
the objects to be freed.

Other FrontEnds (but notably OpenJ9) by default will not free the
options objects, because e.g. OpenJ9 cannot guarantee that the
options objects will not be later read in one of its many event
handlers that continue to operate throughout the shutdown process.

This change addresses crashes in OpenJ9 that started happening
after we plugged the memory leak represented by the options objects
previously not being freed at shutdown.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>