Don't #define VMINLINE to be empty for unoptimized/debug builds
For unoptimized builds with GCC, we only need to avoid the always_inline
attribute. The mere presence of the inline keyword won't cause any
actual inlining (since the build is unoptimized).

As for DEBUG, even if it's defined, many VMINLINE functions are likely
to be inlined in optimized builds anyway, because they're implicitly
declared inline by virtue of being defined within a class declaration.
If a developer wants to prevent inlining for debugging purposes, they
can edit their build flags to disable it (or disable optimization more
generally), or they can inhibit inlining of particular functions in the
source code, e.g. with __attribute__((noinline)).

VMINLINE will now always be defined as VMINLINE_ALWAYS because there is
no prior definition.

This change avoids unused function warnings for standalone static
VMINLINE functions that would end up being declared static but not
inline in unoptimized builds with GCC.

It's still possible for VMINLINE to be empty if VMINLINE_ALWAYS is, but
I ran builds to find when that definition of VMINLINE_ALWAYS is actually
used, i.e. to declare a function, and it only happened in C files on
z/OS and in source generated for DDR.